Made the project compile on macOS, needed to link with correct core frameworks

+++ VulkanProject/src/VulkanSwapChain.cpp
@@ -44,9 +44,9 @@ bool VulkanSwapChain::Init(GLFWwindow* pWindow, VkInstance instance, VkPhysicalD
 		int height = 0;
 		glfwGetWindowSize(pWindow, &width, &height);
 
-		VkExtent2D actualExtent = { width, height };
-		actualExtent.width	= std::max(capabilities.minImageExtent.width, std::min(capabilities.maxImageExtent.width, actualExtent.width));
-		actualExtent.height = std::max(capabilities.minImageExtent.height, std::min(capabilities.maxImageExtent.height, actualExtent.height));
+        VkExtent2D actualExtent = { uint32(width), uint32(height) };
+		actualExtent.width	= std::max(capabilities.minImageExtent.width,   std::min(capabilities.maxImageExtent.width,     actualExtent.width));
+		actualExtent.height = std::max(capabilities.minImageExtent.height,  std::min(capabilities.maxImageExtent.height,    actualExtent.height));
 		m_Extent = actualExtent;
 	}
